sync node-internal-modules-cjs-loader; rename back to old name
