+++ raw/download-and-compare.sh
@@ -3,7 +3,6 @@ set -euo pipefail
 shopt -s inherit_errexit
 __dirname="$(CDPATH= cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 cd "$__dirname"
-set -x
 
 # This script serves as helpful documentation for where these files came from.
 
@@ -14,6 +13,9 @@ function download() {
 compare() {
   diff "$local-$version$ext" "../dist-raw/$local$ext" || true
 }
+assertStrippedIsOnlyDeletions() {
+  [ "$( diff "$1.js" "$1-stripped.js" | grep -E '^>' )" = '' ]
+}
 
 ext=.js
 
@@ -23,15 +25,15 @@ path=lib/internal/modules/cjs/loader
 local=node-internal-modules-cjs-loader
 version=v17.0.1
 download
-compare
+# compare
 
 version=v15.3.0
 download
-compare
+# compare
 
 version=2d5d77306f6dff9110c1f77fefab25f973415770
 download
-compare
+# compare
 
 ####
 
@@ -39,11 +41,11 @@ path=lib/internal/modules/esm/resolve
 local=node-internal-modules-esm-resolve
 version=v13.12.0
 download
-compare
+# compare
 
 version=v15.3.0
 download
-compare
+# compare
 
 ####
 
@@ -51,11 +53,11 @@ path=lib/internal/repl/await
 local=node-internal-repl-await
 version=v17.0.0
 download
-compare
+# compare
 
 version=88799930794045795e8abac874730f9eba7e2300
 download
-compare
+# compare
 
 ####
 
@@ -63,4 +65,12 @@ path=lib/internal/modules/package_json_reader
 local=node-internal-modules-package_json_reader
 version=v15.3.0
 download
-compare
+# compare
+
+####
+
+# Verify that -stripped.js files have only deletions, no other changes
+set -x
+
+assertStrippedIsOnlyDeletions node-internal-modules-cjs-loader-v15.3.0
+assertStrippedIsOnlyDeletions node-internal-modules-cjs-loader-v17.0.1
